Metadata about Project Gutenberg authors

Data frame with metadata about each author of a Project Gutenberg work.

Usage

gutenberg_authors

Format

A tbl_df (see tibble::tibble()) with one row for each author, with the columns

gutenberg_author_id

Unique identifier for the author that can be used to join with the gutenberg_metadata dataset (coming soon to this package)

author

The agent_name field from the original metadata

birthdate

Year of birth

deathdate

Year of death

Details

To find the date on which this metadata was last updated, see attr(gutenberg_authors, "date_updated").

Examples

# date last updated
attr(gutenberg_authors, "date_updated")
#> [1] "2024-09-23"